==> Code/06_generate_tables.ipynb <==
import numpy as np
import pandas as pd

file_path = '../Results/Evaluation/'

df1=pd.read_excel(file_path+'spec1_random_500_20251110_180802.xlsx')
df2=pd.read_excel(file_path+'spec2_recent_500_20251110_181330.xlsx')
df3=pd.read_excel(file_path+'spec3_recent_500_plus_10q_20251110_182010.xlsx')
df4=pd.read_excel(file_path+'spec4_persona_plus_10q_20251110_182428.xlsx')

df1['spec']=1
df2['spec']=2
df3['spec']=3
df4['spec']=4
df_all=pd.concat([df1,df2,df3,df4],ignore_index=True)
df_all.columns

df_all.groupby('spec').agg({'similarity':['mean','std','median']}).round(3)

# Perform the groupby operation
similarity_result = df_all.groupby('spec').agg({'similarity':['mean','std','median']}).round(3)

# Convert the result to LaTeX format
latex_output = similarity_result.to_latex()

# Write the LaTeX string to a .tex file
with open(file_path+'similarity_result.tex', 'w') as tex_file:
    tex_file.write(latex_output)

df_all.groupby(['spec', 'gen_to_gt_label', 'gt_to_gen_label']).size()

df_all.groupby('spec').agg({'tone_alignment': ['mean', 'std', 'median']}).round(3)

# Perform the groupby operation
tone_alignment_result = df_all.groupby('spec').agg({'tone_alignment': ['mean', 'std', 'median']}).round(3)

# Convert the result to LaTeX format
latex_output = tone_alignment_result.to_latex()

# Write the LaTeX string to a .tex file
with open(file_path+'tone_alignment_result.tex', 'w') as tex_file:
    tex_file.write(latex_output)
